Experimental version to get test tokens working with UI
